Put all commands in one file with options, update readme for apiview
