+++ README.rst
@@ -58,14 +58,14 @@ To use DRF Generator, add it your INSTALLED_APPS.
 
 Then run one of the following commands, where ``app`` is the application to generate an API for.
 
-======================== ===================================================
-Command                  Action
-======================== ===================================================
-``generate-serializers`` Generate Serializers for your app.
-``generate-views``       Generate Views for your app.
-``generate-urls``        Generate urls for your app.
-``generate-api``         Generate Serializers, Views, and urls for your app.
-======================== ===================================================
+========================== ===================================================
+Command                    Action
+========================== ===================================================
+``generate --serializers`` Generate Serializers for your app.
+``generate --views``       Generate Views for your app.
+``generate --urls``        Generate urls for your app.
+``generate``               Generate Serializers, Views, and urls for your app.
+========================== ===================================================
 
 .. code-block:: bash
 
@@ -101,7 +101,24 @@ Views
 
 DRF Generator also takes care of all of your basic CRUD API views using your models and the generated serializers.
 
-DRF Generator creates a basic CRUD API View and List View for each model. The basic CRUD view has methods for ``GET``, ``PUT``, and ``DELETE``. The List View has a ``GET`` method that returns a paginated result of the model, and a ``POST`` method to save a new model.
+By default, DRF Generator will create ViewSet View lcasses like the following for your models.
+
+.. code-block:: python
+
+    class CategoryViewSet(ViewSet):
+
+        def list(self, request):
+            ...
+        def create(self, request):
+            ...
+        def retrieve(self, request, pk=None):
+            ...
+        def update(self, request, pk=None):
+            ...
+        def destroy(self, request, pk=None):
+            ...
+
+When running the generator with the ``--apiview`` option, you will get the following API Views.
 
 .. code-block:: python
 
@@ -109,19 +126,15 @@ DRF Generator creates a basic CRUD API View and List View for each model. The ba
 
         def get(self, request, id, format=None):
             ...
-
         def put(self, request, id, format=None):
             ...
-
         def delete(self, request, id, format=None):
             ...
-
 
     class UserAPIListView(APIView):
 
         def get(self, request, format=None):
             ...
-
         def post(self, request, format=None):
             ...
 
@@ -131,6 +144,18 @@ Urls
 ----
 
 Finally, DRF Generator will create you a default ``urls.py`` in the following format.
+
+By default, DRF Generator will create rouserce route based urls like the following.
+
+.. code-block:: python
+
+    router = SimpleRouter()
+
+    router.register(r'model', views.ModelViewSet, 'Model')
+
+    urlpatterns = router.urls
+
+If you run the generatro with the ``--apiview`` option, you will get urls like the following.
 
 .. code-block:: python
