--- drf_generators/management/commands/generate-serializers.py
@@ -1,17 +1,0 @@
-
-from django.core.management.base import AppCommand
-from drf_generators.generators import BaseGenerator
-
-
-class Command(AppCommand):
-    help = 'Generates DRF API Views and Serializers for a Django app'
-
-    args = "[appname ...]"
-
-    def handle_app_config(self, app_config, **options):
-        if app_config.models_module is None:
-            raise CommandError('You must provide an app to generate an API')
-
-        generator = BaseGenerator(app_config)
-        result = generator.generate_serializers()
-        print(result)
